+++ src/com/devstackio/fullstackcoder/entity/enemy/EnemyBlock.java
@@ -6,23 +6,48 @@ import org.newdawn.slick.Graphics;
 
 /**
  * Block of enemies - contains List of Enemy entities that will match to active CodeBlock
- * @author devstackio <@devstackioweb>
+ * @author devstackio @devstackioweb
  */
 public class EnemyBlock {
     
     private static Graphics GRAPHIX;
+    private EnemyFactory enemyFactory;
     private float speed=.5f;
     private List<EnemyMold> enemies;
+    private int enemyCount;
     
-    private List<EnemyMold> createCodeLines( String[] lines ) {
+    public EnemyBlock( Graphics g, int num ) {
+        GRAPHIX = g;
+        this.enemyCount = num;
+        this.enemyFactory = new EnemyFactory();
+        this.enemies = this.createEnemies();
+    }
+    
+    private List<EnemyMold> createEnemies() {
         List<EnemyMold> returnobj = new LinkedList();
-        int counter=0;
-        for ( String line : lines ) {
-//            EnemyMold codeLine = new FlyingEnemy();
-//            returnobj.add( codeLine );
-//            counter++;
+        
+        for (int i = 0; i < this.enemyCount; i++) {
+            
+            returnobj.add( this.enemyFactory.createRandomEnemy() );
+            
         }
         return returnobj;
+    }
+
+    public List<EnemyMold> getEnemies() {
+        return enemies;
+    }
+    
+    public void ioDraw() {
+        for ( EnemyMold enemy : this.enemies ) {
+            enemy.draw();
+        }
+    }
+    
+    public void ioUpdate( int delta ) {
+        for ( EnemyMold enemy : this.enemies ) {
+            enemy.update( delta );
+        }
     }
     
 }