+++ src/com/devstackio/fullstackcoder/entity/enemy/EnemyMold.java
@@ -2,7 +2,7 @@ package com.devstackio.fullstackcoder.entity.enemy;
 
 /**
  * creating a basic Enemy interface since we should be able to swap between different types of enemies later ( flying, walking, whatever else comes up )
- * @author devstackio <@devstackioweb>
+ * @author devstackio @devstackioweb
  */
 public interface EnemyMold {