+++ src/com/devstackio/fullstackcoder/observer/ActionObserver.java
@@ -1,0 +1,7 @@
+package com.devstackio.fullstackcoder.observer;
+
+public interface ActionObserver {
+       
+    public void update( ActionType actionType );
+    
+}