+++ src/com/devstackio/fullstackcoder/observer/ActionType.java
@@ -1,0 +1,7 @@
+package com.devstackio.fullstackcoder.observer;
+
+public enum ActionType {
+    
+    LINE_COMPLETE,BLOCK_COMPLETE,LEVEL_COMPLETE;
+    
+}